requirements.txt
anyio==4.6.2.post1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==13.13
pytz==2024.2
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
tzlocal==5.2
urllib3==1.26.5
